Large progress on game service and models

+++ backend/src/main/java/org/dev/commander/model/game/Space.java
@@ -1,22 +1,22 @@
 package org.dev.commander.model.game;
 
 public class Space {
-    private int width;
-    private int height;
+    private double width;
+    private double height;
 
-    public int getWidth() {
+    public double getWidth() {
         return width;
     }
 
-    public void setWidth(int width) {
+    public void setWidth(double width) {
         this.width = width;
     }
 
-    public int getHeight() {
+    public double getHeight() {
         return height;
     }
 
-    public void setHeight(int height) {
+    public void setHeight(double height) {
         this.height = height;
     }
 }
